+++ weblate/static/loader-bootstrap.js
@@ -1681,7 +1681,10 @@ $(function () {
           .find(".file-format-param-field")
           .attr("fileformats")
           ?.split(" ");
-        if (fileFormats.includes(selectedFileFormat)) {
+        if (
+          fileFormats.includes(selectedFileFormat) ||
+          fileFormats.includes("*")
+        ) {
           $(this).show();
           displayFieldLabel = true;
         } else {